% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_rest.R
\name{mle_rest}
\alias{mle_rest}
\title{Maximum likelihood parameter estimation for the REST/RAD-REST model based on the implementation of Nelder and Mead (1965).}
\usage{
mle_rest(station_effort_data, stay_data, activity_data, focal_area, model)
}
\arguments{
\item{station_effort_data}{The output of the add_effort function.}

\item{stay_data}{The output of the format_stay function.}

\item{activity_data}{The output of the format_activity function.}

\item{focal_area}{The area of the focal area.}

\item{model}{Specify the model to use. Choose either "REST" or "RAD-REST".}
}
\value{
A data frame containing the estimated density, its standard deviation (SD), the 95\% lower and upper confidence limits, AIC, and delta AIC value for each model.
}
\description{
Maximum likelihood parameter estimation for the REST/RAD-REST model based on the implementation of Nelder and Mead (1965).
}
\examples{
station_data_rest <- format_station_data(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_y = "y",
  model = "REST",
  target_species = "SP01"
)

station_effort_data <- add_effort(
  detection_data = detection_data,
  station_data_formatted = station_data_rest,
  col_name_station = "Station",
  col_name_term = "Term",
  col_name_datetime = "DateTime",
  plot = TRUE,
  font_size = 5
)

stay_data <- format_stay(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens",
  target_species = "SP01"
)

activity_data <- format_activity(
  detection_data = detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_datetime = "DateTime",
  target_species = "SP01",
  indep_time = 30
)

model <- mle_rest(station_effort_data,
                  stay_data,
                  activity_data,
                  focal_area = 1.56,
                  model = "REST")

}
