% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effort.R
\name{add_effort}
\alias{add_effort}
\title{Calculate Camera Trapping Effort (Days) from Detection Data}
\usage{
add_effort(
  detection_data,
  station_data_formatted,
  col_name_station,
  col_name_datetime,
  col_name_term = NULL,
  plot = FALSE
)
}
\arguments{
\item{detection_data}{A data frame containing information on individual video records.}

\item{station_data_formatted}{A data frame returned by the \code{format_station_data()} function.}

\item{col_name_station}{A string specifying the column name containing camera station IDs.}

\item{col_name_datetime}{A string specifying the column name containing datetime information.}

\item{col_name_term}{(Optional) A string specifying the column name indicating different survey periods.}

\item{plot}{Logical. If TRUE, a plot showing the camera operation periods will be displayed.}
}
\value{
A data frame with an additional column \code{Effort}. Stations with \code{Effort == 0} or no detections are removed.
}
\description{
Computes the total camera trapping effort (in days) for each station based on detection data.
}
