% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_effort.R
\name{add_effort}
\alias{add_effort}
\title{Caluculating camera trapping efforts from the detection data}
\usage{
add_effort(
  detection_data,
  npass_data,
  col_name_station = "station",
  col_name_datetime = "datetime"
)
}
\arguments{
\item{detection_data}{A data frame containing information for individual videos, with six columns: camera station name (station), capture datetime (datetime), species name (species), number of passes through the focal area (nfocal), duration of stay within the focal area for each pass in a video (stay), and whether the observation of the pass was censored (cens).}

\item{npass_data}{""}

\item{col_name_station}{Column name containing station id info}

\item{col_name_datetime}{Column name containing datetime info}
}
\value{
data frame containing camera trapping effort (days) for each station
}
\description{
Caluculating camera trapping efforts from the detection data
}
\details{
'start' and 'end' represent the dates when the camera trap first and last captured an image, respectively.
}
\examples{
add_effort(detection_data = detection_data,
npass_data = station_data,
col_name_station = "station",
col_name_datetime = "datetime")
}
