% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_activity.R
\name{bayes_activity}
\alias{bayes_activity}
\title{Activity level estimation by von Misex mixture distribution using stick-breaking prior}
\usage{
bayes_activity(
  activity_data = activity_data,
  C = 10,
  plot = TRUE,
  cores = 3,
  iter = 3000,
  warmup = 1000,
  chains = 3,
  thin = 5,
  target_species = NULL
)
}
\arguments{
\item{activity_data}{A vector of detection times transformed into radians}

\item{C}{""}

\item{plot}{If TRUE, plots the expected values of residence times.}

\item{cores}{The number of cores used for calculations.}

\item{iter}{The number of iterations. The default value is 2000.}

\item{warmup}{The number of warmup iterations. The default value is NULL (half of the iterations will be used for warmup).}

\item{chains}{The number of chains. The default value is 2.}

\item{thin}{The thinning interval. The default value is 1.}

\item{target_species}{Species name of interest.}

\item{all_comb}{If TRUE, models with all combinations of covariates are compared. If FALSE, only the designated model in model_formula is run.}
}
\value{
Activity level
}
\description{
Activity level estimation by von Misex mixture distribution using stick-breaking prior
}
\examples{
activity_data <- format_activity(
  detection_data = detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_datetime = "DateTime",
  target_species = "SP01",
  indep_time = 30
)
library(nimble)
bayes_activity(
  activity_data = activity_data,
  C = 10,
  plot = TRUE,
  iter = 2000,
  warmup = 1000,
  chains = 2,
  thin = 1,
  all_comb = FALSE,
  target_species = "SP01")
}
