% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_activity.R
\name{bayes_activity}
\alias{bayes_activity}
\title{Activity level estimation by von Mises mixture distribution using stick-breaking prior}
\usage{
bayes_activity(
  activity_data,
  C = 10,
  cores = 3,
  iter = 5000,
  warmup = 1000,
  chains = 3,
  thin = 2,
  target_species
)
}
\arguments{
\item{activity_data}{A data frame containing a "time" column, representing detection times transformed into radians. Typically, this is the output of the format_activity function.}

\item{C}{The maximum number of components in the von Mises mixture distribution. Defaults to 10.}

\item{cores}{The number of CPU cores to use for parallel computation. Default is 3.}

\item{iter}{The total number of MCMC iterations per chain. Default is 5000}

\item{warmup}{The number of warm-up (burn-in) iterations per chain. Default is 1000.}

\item{chains}{The number of MCMC chains. Default is 2.}

\item{thin}{The thinning interval for MCMC sampling. Default is 4 (no thinning).}

\item{target_species}{The species name of interest.  Only one species can be specified.}
}
\value{
A list containing:
\describe{
\item{summary_result}{A data frame summarizing the estimated activity proportion.}
\item{plot}{A ggplot2 object showing the temporal variation in activity level.}
}
}
\description{
Activity level estimation by von Mises mixture distribution using stick-breaking prior
}
\examples{
activity_data <- format_activity(
  detection_data = detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_datetime = "DateTime",
  indep_time = 30
)
bayes_activity(
  activity_data = activity_data,
  C = 10,
  cores = 1,
  iter = 5000,
  warmup = 1000,
  chains = 1,
  thin = 4,
  target_species = "SP01")
}
