% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_rest.R
\name{bayes_rest}
\alias{bayes_rest}
\title{Bayesian parameter estimation of REST/RAD-REST model based on MCMC samplings with Rstan}
\usage{
bayes_rest(
  formula_stay,
  formula_density,
  station_data_complete,
  stay_data,
  activity_data,
  activity_model = "kernel",
  K = 5,
  bw_adj = 1.5,
  stay_family = "lognormal",
  local_stay = FALSE,
  focal_area,
  cores = 2,
  iter = 2000,
  warmup = NULL,
  chains = 2,
  thin = 1,
  model = "REST"
)
}
\arguments{
\item{formula_stay}{Model formula for the staying time within a focal area. Grammar follows lme4::glmer function. e.g. Stay ~ 1 + (1|Group)}

\item{formula_density}{Model formula for animal density. e.g. ~ 1 + x1}

\item{station_data_complete}{A data frame containing information for each camera station in each row}

\item{stay_data}{A data frame containing the time spent within a focal area}

\item{activity_data}{A vector of detection times transformed into radians}

\item{activity_model}{The method for estimating the proportion of active animals. If "kernel", the fixed kernel method by Rowcliffe et al. () is used. If "mixed", a von Mises mixture distribution is used. The default value is "kernel".}

\item{K}{The number of mixtures when activity_model = "mixed". The default value is 5.}

\item{bw_adj}{Bandwidth adjustment when activity_model = "mixed". The default value is 1.5. See Rowcliffe et al. () for details.}

\item{stay_family}{The probability distribution of the time spent. Specify the probability distribution selected by the bayes_stay_selection function.}

\item{local_stay}{Specify whether to use the expected value of the time spent for each camera station (local_stay = TRUE) or the global expected value (FALSE) for density estimation.}

\item{focal_area}{The area of a focal area}

\item{cores}{The number of cores to use for parallel computation. The default value is 1.}

\item{iter}{The length of iterations. The default value is 2000.}

\item{warmup}{The length of warmup. The default value is NULL (half of the iterations will be used for warmup).}

\item{chains}{The number of chains. The default value is 2.}

\item{thin}{The thinning interval. The default value is 1.}

\item{model}{Specify the model to use. Choose either "REST" or "RAD-REST".}
}
\value{
An object of class "stanfit" representing
the fitted Bayesian model. It contains samples from the posterior
distribution of the model parameters, which can be accessed using
the rstan extraction functions like rstan::extract. The object also
contains other useful quantities like the MCMC chains, sampling metadata etc.
}
\description{
Bayesian parameter estimation of REST/RAD-REST model based on MCMC samplings with Rstan
}
\examples{
station_data_rest <- format_station_data(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_y = "y",
  model = "REST",
  target_species = "A"
)

station_data_complete <- add_effort(
  detection_data = detection_data,
  station_data_formatted = station_data_rest,
  col_name_station = "Station",
  col_name_term = "Term",
  col_name_datetime = "DateTime",
  plot = TRUE,
  font_size = 5
)

stay_data <- format_stay(
  detection_data = detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens",
  target_species = "A"
)

activity_data <- format_activity(
  detection_data = detection_data,
  col_name_species = "Species",
  col_name_datetime = "DateTime",
  target_species = "A"
)

fitstan <- bayes_rest(formula_stay = Stay ~ 1,
  formula_density = ~ 1,
  station_data_complete,
  stay_data,
  activity_data,
  activity_model = "kernel",
  local_stay = FALSE,
  stay_family = "exponential",
  focal_area = 2.0,
  model = "REST")
}
