% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_rest.R
\name{bayes_rest}
\alias{bayes_rest}
\title{Bayesian parameter estimation for the REST/RAD-REST model via MCMC sampling using \code{nimble}.}
\usage{
bayes_rest(
  formula_stay,
  formula_density,
  station_effort_data,
  stay_data,
  random_effect = NULL,
  activity_data,
  activity_estimation = "kernel",
  C = 10,
  bw_adj = 1,
  stay_family = "lognormal",
  focal_area,
  cores = 3,
  iter = 5000,
  warmup = 1000,
  chains = 3,
  thin = 2,
  model = "REST",
  all_comb = FALSE,
  target_species
)
}
\arguments{
\item{formula_stay}{A model formula for staying times within the focal area. For example, \code{Stay ~ 1 + x1}.}

\item{formula_density}{A model formula for animal density. For example, \code{~ 1 + x1} (note that the left-hand side must be omitted).}

\item{station_effort_data}{A data frame containing information for each camera station. Typically, this is the output of the \code{add_effort} function. Alternatively, a manually prepared data frame may be provided, with the required columns as follows:
\itemize{
\item If \code{model = "REST"}, the data frame must contain:
\itemize{
\item \code{Station} (character): Camera station ID.
\item \code{Effort} (numeric): Camera trapping effort (days) for each station.
\item \code{Species} (character): Species name.
\item \code{Y} (numeric): Total number of passes through the focal area for each station-species pair.
}
\item If \code{model = "RAD-REST"}, the data frame must contain:
\itemize{
\item \code{Station} (character): Camera station ID.
\item \code{Effort} (numeric): Camera trapping effort (days) for each station.
\item \code{Species} (character): Species name.
\item \code{N} (numeric): Total number of detected videos for each station-species pair.
\item \code{y_X} columns (\code{y_0}, \code{y_1}, ..., \code{y_max}): Number of videos classified by the number of passes observed.
}
}}

\item{stay_data}{A data frame returned by the \code{format_stay} function, containing:
\itemize{
\item \code{Station} (character): Camera station ID.
\item \code{Species} (character): Species name.
\item \code{Stay} (numeric): Staying time (in seconds) within the focal area for each detected pass.
\item \code{Censored} (binary): Whether the staying time was censored (1 = censored, 0 = observed).
}}

\item{random_effect}{A specification of the random effect structure on staying time (e.g., \code{~ (1 | Station)}).}

\item{activity_data}{A data frame containing a "time" column, representing detection times transformed into radians. Typically, this is the output of the format_activity function.}

\item{activity_estimation}{Method used to estimate activity patterns. Choose \code{"kernel"} for fixed kernel density estimation (Rowcliffe et al. 2014) or \code{"mixture"} for nonparametric Bayesian estimation using von Mises mixture models (Nakashima et al. 2025).}

\item{C}{The maximum number of von Mises components used for the mixture model. Required only if \code{activity_estimation = "mixture"}. The default is 10.}

\item{bw_adj}{Bandwidth adjustment parameter for kernel density estimation. The default is 1.0. See Rowcliffe et al. (2014) for details.}

\item{stay_family}{The probability distribution used for modeling staying times. This should be selected based on the output of the \code{bayes_stay_selection} function.}

\item{focal_area}{The area of the focal area, in square meters.}

\item{cores}{The number of CPU cores to use for parallel computation. Default is 3.}

\item{iter}{The total number of MCMC iterations per chain. Default is 5000}

\item{warmup}{The number of warm-up (burn-in) iterations per chain. Default is 1000.}

\item{chains}{The number of MCMC chains. Default is 2.}

\item{thin}{The thinning interval for MCMC sampling. Default is 1 (no thinning).}

\item{model}{The model to be used. Choose either \code{"REST"} or \code{"RAD-REST"}.}

\item{all_comb}{Logical. If TRUE, all combinations of covariates in the density model will be evaluated and compared. If FALSE, only the specified model in \code{formula_density} will be run.}

\item{target_species}{A character string specifying the species to be analyzed. Only a single species can be specified.}
}
\value{
A list containing:
}
\description{
Bayesian parameter estimation for the REST/RAD-REST model via MCMC sampling using \code{nimble}.
}
\examples{
station_data_rest <- format_station_data(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_y = "y",
  model = "REST",
  target_species = "SP01"
)
station_effort_REST <- add_effort(
  detection_data = detection_data,
  station_data_formatted = station_data_rest,
  col_name_station = "Station",
  col_name_term = "Term",
  col_name_datetime = "DateTime",
  plot = TRUE,
  font_size = 5
)
stay_data <- format_stay(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens",
  target_species = "SP01"
)
activity_data <- format_activity(
  detection_data = detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_datetime = "DateTime",
  target_species = "SP01",
  indep_time = 30
)
rest_model <- bayes_rest(
  formula_stay = Stay ~ 1 + x1,
  formula_density = ~ 1,
  station_effort_data = station_effort_REST,
  stay_data = stay_data,
  activity_data = activity_data,
  activity_estimation = "kernel",
  bw_adj = 1.5,
  stay_family = "lognormal",
  focal_area = 1.96,
  cores = 2,
  iter = 3000,
  warmup = 1000,
  chains = 2,
  thin = 1,
  model = "REST",
  all_comb = FALSE,
  target_species = "SP01"
)

}
