% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_rest.R
\name{bayes_rest}
\alias{bayes_rest}
\title{Bayesian parameter estimation of REST/RAD-REST model based on MCMC samplings with Rstan}
\usage{
bayes_rest(
  formula_stay,
  formula_density,
  station_effort_data,
  stay_data,
  col_name_cens = "Cens",
  random_effect = NULL,
  activity_data,
  bw_adj = 1.5,
  stay_family = "lognormal",
  focal_area,
  cores = 2,
  iter = 3000,
  warmup = NULL,
  chains = 2,
  thin = 1,
  model = "REST",
  all_comb = F
)
}
\arguments{
\item{formula_stay}{Model formula for the staying time within a focal area. Grammar follows lme4::glmer function. e.g. Stay ~ 1 + (1|Group)}

\item{formula_density}{Model formula for animal density. e.g. ~ 1 + x1}

\item{station_effort_data}{A data frame containing information for each camera station (with sampling days) in each row}

\item{stay_data}{A data frame containing the time spent within a focal area}

\item{col_name_cens}{Column name of stay_data containing information on censored (1) or not (0)}

\item{random_effect}{A random effect on mean staying time}

\item{activity_data}{A vector of detection times transformed into radians}

\item{bw_adj}{Bandwidth adjustment of kernel density estimation. The default value is 1.5. See Rowcliffe et al. () for details.}

\item{stay_family}{The probability distribution of staying times. Specify the probability distribution selected by the bayes_stay_selection function.}

\item{focal_area}{The area of a focal area}

\item{cores}{The number of cores to use for parallel computation. The default value is 1.}

\item{iter}{The length of iterations. The default value is 2000.}

\item{warmup}{The length of warmup. The default value is NULL (half of the iterations will be used for warmup).}

\item{chains}{The number of chains. The default value is 2.}

\item{thin}{The thinning interval. The default value is 1.}

\item{model}{Specify the model to use. Choose either "REST" or "RAD-REST".}

\item{all_comb}{If TRUE, models with all combinations of covariates are compared. If FALSE, only the designated model in model_formula is run.}
}
\value{
XXXX
}
\description{
Bayesian parameter estimation of REST/RAD-REST model based on MCMC samplings with Rstan
}
\examples{
station_data_rest <- format_station_data(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_y = "y",
  model = "REST",
  target_species = "SP01"
)
station_effort_REST <- add_effort(
  detection_data = detection_data,
  station_data_formatted = station_data_rest,
  col_name_station = "Station",
  col_name_term = "Term",
  col_name_datetime = "DateTime",
  plot = TRUE,
  font_size = 5
)
stay_data <- format_stay(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens",
  target_species = "SP01"
)
activity_data <- format_activity(
  detection_data = detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_datetime = "DateTime",
  target_species = "SP01",
  indep_time = 30
)
library(nimble)
rest_model <- bayes_rest(
  formula_stay = Stay ~ 1 + x1,
  formula_density = ~ 1,
  station_effort_data = station_effort_REST,
  stay_data = stay_data,
  col_name_cens = "Cens",
  activity_data = activity_data,
  bw_adj = 1.5,
  stay_family = "lognormal",
  focal_area = 1.96,
  cores = 2,
  iter = 3000,
  warmup = 1000,
  chains = 2,
  thin = 1,
  model = "REST",
  all_comb = FALSE
)

}
