% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_rest_multi.R
\name{bayes_rest_multi}
\alias{bayes_rest_multi}
\title{Bayesian Parameter Estimation of a Multispecies RAD-REST Model Based on MCMC Sampling with RStan}
\usage{
bayes_rest_multi(
  formula_stay,
  formula_density,
  station_effort_data,
  stay_data,
  random_effect = NULL,
  activity_data,
  activity_estimation = "kernel",
  bw_adj = 1,
  C = 10,
  stay_family = "lognormal",
  focal_area,
  cores = 3,
  iter = 5000,
  warmup = 1000,
  chains = 3,
  thin = 4,
  target_species,
  all_comb = FALSE
)
}
\arguments{
\item{formula_stay}{A model formula for staying times within the focal area (e.g., Stay ~ 1 + x1). Random effects should be specified separately using the random_effect argument.}

\item{formula_density}{A model formula for animal density (e.g., ~ 1 + x1). Note that the formula must not have a left-hand side.}

\item{station_effort_data}{A data frame containing information for each camera station. Typically, this is the output of the add_effort function. Alternatively, a manually prepared data frame may be provided with the following required columns:
\itemize{
\item For model = "REST":
\item Station (character): Camera station ID.
\item Effort (numeric): Camera trapping effort (in days) at each station.
\item Species (character): Species name.
\item Y (numeric): Total number of passes through the focal area for each station-species pair.
\item For model = "RAD-REST":
\item Station (character): Camera station ID.
\item Effort (numeric): Camera trapping effort (in days) at each station.
\item Species (character): Species name.
\item N (numeric): Total number of detected videos for each station-species pair.
\item y_X columns (y_0, y_1, ..., y_max): Number of videos categorized by the number of observed passes.
}}

\item{stay_data}{A data frame returned by the format_stay function, containing:
\itemize{
\item Station (character): Camera station ID.
\item Species (character): Species name.
\item Stay (numeric): Staying time (in seconds) within the focal area for each detected pass.
\item Censored (binary): Whether the staying time was censored (1 = censored, 0 = observed).
}}

\item{random_effect}{A character string specifying a random effect on mean staying time. (Note: Species-level random effects are automatically included and do not need to be specified.)}

\item{activity_data}{A data frame containing a "time" column, representing detection times transformed into radians. Typically, this is the output of the format_activity function.}

\item{activity_estimation}{The method used to estimate activity patterns. Choose "kernel" for fixed kernel density estimation (Rowcliffe et al. 2014), or "mixture" for nonparametric Bayesian estimation using von Mises mixture models (Nakashima et al. 2025).}

\item{bw_adj}{Bandwidth adjustment parameter for kernel density estimation. The default is 1.0. See Rowcliffe et al. (2014) for details.}

\item{C}{The maximum number of von Mises components to use in the mixture model. Required only if activity_estimation = "mixture".}

\item{stay_family}{The probability distribution used to model staying times. This should be selected based on the output of the bayes_stay_selection function.}

\item{focal_area}{The size of the focal area, in square meters.}

\item{cores}{The number of CPU cores to use for parallel computation. Default is 3.}

\item{iter}{The total number of MCMC iterations per chain. Default is 5,000.}

\item{warmup}{The number of warm-up (burn-in) iterations per chain. Default is 1,000.}

\item{chains}{The number of MCMC chains. Default is 3.}

\item{thin}{The thinning interval for MCMC sampling. Default is 4 (no thinning).}

\item{target_species}{A vector specifying the species to be analyzed. Multiple species must be specified.}

\item{all_comb}{Logical. If TRUE, all combinations of covariates in the density model will be evaluated and compared. If FALSE, only the model specified in formula_density will be run.}
}
\value{
A list of class \code{"ResultDensity"}, which includes the following components:
\describe{
\item{\code{WAIC}}{An object containing WAIC (Widely Applicable Information Criterion) results.}
\item{\code{summary_result}}{A data frame summarizing the posterior estimates of the mean staying time.}
\item{\code{samples}}{A \code{coda::mcmc.list} object containing MCMC samples for all model parameters.}
}
The returned object has a custom print method that displays the WAIC and summary of staying time estimates.
You can access full MCMC samples via \code{$samples}, and analyze convergence using the \code{MCMCvis} package.
}
\description{
Bayesian Parameter Estimation of a Multispecies RAD-REST Model Based on MCMC Sampling with RStan
}
\examples{
station_data_RAD <- format_station_data(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_y = "y",
  model = "RAD-REST"
)
station_effort_RAD <- add_effort(
  detection_data = detection_data,
  station_data_formatted = station_data_RAD,
  col_name_station = "Station",
  col_name_term = "Term",
  col_name_datetime = "DateTime",
  plot = TRUE
)
stay_data <- format_stay(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens"
)
activity_data <- format_activity(
  detection_data = detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_datetime = "DateTime",
  indep_time = 30
)
rest_model <- bayes_rest_multi(
  formula_stay = Stay ~ 1,
  formula_density = ~ 1,
  station_effort_data = station_effort_RAD,
  stay_data = stay_data,
  activity_data = activity_data,
  stay_family = "lognormal",
  focal_area = 1.96,
  cores = 2,
  iter = 5000,
  warmup = 1000,
  chains = 2,
  thin = 1,
  target_species = c("SP01", "SP02", "SP03", "SP04", "SP05"),
  all_comb = FALSE
)

}
