% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_stay_selection.R
\name{bayes_stay_selection}
\alias{bayes_stay_selection}
\title{Model selection for the residence time model based on WAIC for REST/REST-RAD models}
\usage{
bayes_stay_selection(
  formula_stay,
  station_data,
  stay_data,
  family = "lognormal",
  local_stay = FALSE,
  plot = TRUE,
  cores = 3,
  iter = 2000,
  warmup = NULL,
  chains = 2,
  thin = 1,
  all_comb = TRUE
)
}
\arguments{
\item{formula_stay}{Model formula for the residence time within a focal area. Grammar follows lme4::glmer function. e.g. Stay ~ 1 + (1|Group)}

\item{station_data}{A data frame containing information for each camera station in each row}

\item{stay_data}{A data frame containing the time spent within a focal area}

\item{family}{The probability distribution of the time spent. Specify the probability distribution selected by the bayes_stay_selection function.}

\item{local_stay}{Specify whether to use the expected value of the time spent for each camera station (local_stay = TRUE) or the global expected value (FALSE) for density estimation.}

\item{plot}{If TRUE, plots the expected values of residence times.}

\item{cores}{The number of cores to use for parallel computation. The default value is 1.}

\item{iter}{The number of iterations. The default value is 2000.}

\item{warmup}{The number of warmup iterations. The default value is NULL (half of the iterations will be used for warmup).}

\item{chains}{The number of chains. The default value is 2.}

\item{thin}{The thinning interval. The default value is 1.}

\item{all_comb}{If TRUE, models with all combinations of covariates are compared. If FALSE, the WAIC value of the designated model is shown.}
}
\value{
WAIC values for each model
}
\description{
Model selection for the residence time model based on WAIC for REST/REST-RAD models
}
\examples{
stay_data <- format_stay(
  detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens",
  target_species = "A"
)
bayes_stay_selection(formula_stay = Stay ~ 1 + x1,
                    station_data,
                    stay_data,
                    family = "lognormal",
                    local_stay = TRUE,
                    plot = TRUE,
                    cores = 3,
                    iter = 2000,
                    warmup = NULL,
                    chains = 2,
                    thin = 1,
                    all_comb = FALSE)
}
