% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_stay_selection.R
\name{bayes_stay_selection}
\alias{bayes_stay_selection}
\title{Model selection for the residence time model based on WAIC for REST/REST-RAD models}
\usage{
bayes_stay_selection(
  formula_stay = Stay ~ 1,
  random_effect = NULL,
  station_data = station_data,
  stay_data = stay_data,
  col_name_cens = "Cens",
  family = "lognormal",
  plot = TRUE,
  cores = 1,
  iter = 3000,
  warmup = NULL,
  chains = 3,
  thin = 1,
  all_comb = TRUE,
  target_species = NULL
)
}
\arguments{
\item{formula_stay}{Model formula for the residence time within a focal area. e.g. Stay ~ 1 + x1}

\item{random_effect}{random_effect effect on the parameter of staying time model (as a character). e.g. random_effect = "Station"}

\item{station_data}{A data frame containing information for each camera station in each row}

\item{stay_data}{A data frame containing the time spent within a focal area}

\item{col_name_cens}{Column name of stay_data containing information on censored (1) or not (0)}

\item{family}{The probability distribution of the time spent. Specify the probability distribution selected by the bayes_stay_selection function.}

\item{plot}{If TRUE, plots the expected values of residence times.}

\item{cores}{The number of cores used for calculations.}

\item{iter}{The number of iterations. The default value is 2000.}

\item{warmup}{The number of warmup iterations. The default value is NULL (half of the iterations will be used for warmup).}

\item{chains}{The number of chains. The default value is 2.}

\item{thin}{The thinning interval. The default value is 1.}

\item{all_comb}{If TRUE, models with all combinations of covariates are compared. If FALSE, only the designated model in model_formula is run.}

\item{target_species}{Species name of interest.}
}
\value{
WAIC values for each model
}
\description{
Model selection for the residence time model based on WAIC for REST/REST-RAD models
}
\examples{
stay_data <- format_stay(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens",
  target_species = "SP01"
)
library(nimble)
bayes_stay_selection(formula_stay = Stay ~ 1 + x1,
                    station_data = station_data,
                    stay_data = stay_data,
                    col_name_cens = "Cens",
                    family = "lognormal",
                    plot = TRUE,
                    iter = 2000,
                    warmup = 1000,
                    chains = 2,
                    thin = 1,
                    all_comb = FALSE,
                    target_species = "SP01")
}
