% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_station_data.R
\name{format_station_data}
\alias{format_station_data}
\title{Aggregate the number of animal passes within focal areas for each station from detection data.}
\usage{
format_station_data(
  detection_data,
  station_data,
  col_name_station,
  col_name_species,
  col_name_y,
  model
)
}
\arguments{
\item{detection_data}{A data frame containing information for individual videos.}

\item{station_data}{A data frame containing information for each camera station.}

\item{col_name_station}{A string specifying the column name containing station ID information.}

\item{col_name_species}{A string specifying the column name containing detected species names.}

\item{col_name_y}{A string specifying the column name containing the number of animal passes.}

\item{model}{A string specifying the model type, either "REST" or "RAD-REST".}
}
\value{
A data frame with aggregated detection counts joined with station data.
}
\description{
Aggregate the number of animal passes within focal areas for each station from detection data.
}
