% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_station_data.R
\name{format_station_data}
\alias{format_station_data}
\title{Aggregate the number of passes within focal areas for each station from the detection data}
\usage{
format_station_data(
  detection_data,
  station_data,
  col_name_station,
  col_name_species,
  col_name_y,
  target_species,
  model
)
}
\arguments{
\item{detection_data}{A data frame containing information for individual videos, with six columns: camera station name (station), capture datetime (datetime), species name (species), number of passes through the focal area (nfocal), duration of stay within the focal area for each pass in a video (stay), and whether the observation of the pass was censored (cens).}

\item{station_data}{A data frame containing information for each camera station, with one row per station. It includes the station name (station), location coordinates, covariates, and so forth. Unlike the detection_data, it lists all the camera stations that have been set up.}

\item{col_name_station}{Column name containing station id info}

\item{col_name_species}{Column name containing species name detected}

\item{col_name_y}{Column name containing number of animal passage within a focal area}

\item{target_species}{The species name for which you want to estimate his density}

\item{model}{Model name used ("REST" or "RAD-REST")}
}
\value{
data frame containing at least 3 columns (station, species and npass)
}
\description{
Aggregate the number of passes within focal areas for each station from the detection data
}
\examples{
format_station_data(
  detection_data = detection_data,
  station_data = station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_y = "y",
  model = "REST",
  target_species = "SP01"
)
}
