% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stay.R
\name{format_stay}
\alias{format_stay}
\title{Prepares data for analyzing animal staying time within the focal area}
\usage{
format_stay(
  detection_data,
  station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens"
)
}
\arguments{
\item{detection_data}{A data frame containing information for individual videos.}

\item{station_data}{A data frame containing information for each camera station.}

\item{col_name_station}{String. Column name for station ID.}

\item{col_name_species}{String. Column name for species names.}

\item{col_name_stay}{String. Column name for staying time (numeric).}

\item{col_name_cens}{String. Column name for censoring indicator (0 or 1).}
}
\value{
A joined data frame (tibble).
}
\description{
Prepares data for analyzing animal staying time within the focal area
}
