% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stay.R
\name{format_stay}
\alias{format_stay}
\title{Pre}
\usage{
format_stay(
  detection_data,
  station_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens",
  target_species = "A"
)
}
\arguments{
\item{detection_data}{A data frame containing information for each camera station, with one row per station. It includes the station name (station), location coordinates, covariates, and so forth. Unlike the detection_data, it lists all the camera stations that have been set up.}

\item{station_data}{""}

\item{col_name_station}{Column name containing station id info}

\item{col_name_species}{Column name containing species name detected}

\item{col_name_stay}{Column name containing staying time}

\item{col_name_cens}{Column name containing infomation on censored (1) or not (0)}

\item{target_species}{The species name for which you want to estimate his density}
}
\value{
Dataframe contaning staying time
}
\description{
Pre
}
\examples{
format_stay(
 detection_data = detection_data,
 station_data = station_data,
 col_name_station = "Station",
 col_name_species = "Species",
 col_name_stay = "Stay",
 col_name_cens = "Cens",
 target_species = "SP01")
}
