% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stay.R
\name{format_stay}
\alias{format_stay}
\title{Prepare staying}
\usage{
format_stay(
  detection_data,
  col_name_station = "Station",
  col_name_species = "Species",
  col_name_stay = "Stay",
  col_name_cens = "Cens",
  target_species = "A"
)
}
\arguments{
\item{detection_data}{A data frame containing information for individual videos, with six columns: camera station name (station), capture datetime (datetime), species name (species), number of passes through the focal area (nfocal), duration of stay within the focal area for each pass in a video (stay), and whether the observation of the pass was censored (cens).}

\item{col_name_station}{Column name containing station id info}

\item{col_name_species}{Column name containing species name detected}

\item{col_name_stay}{Column name containing staying time}

\item{col_name_cens}{Column name containing infomation on censored (1) or not (0)}

\item{target_species}{The species name for which you want to estimate his density}
}
\value{
Dataframe contaning staying time
}
\description{
Prepare staying
}
\examples{
format_stay(detection_data = detection_data,
 col_name_station = "Station",
 col_name_species = "Species",
 col_name_stay = "Stay",
 col_name_cens = "Cens",
 target_species = "A")
}
