% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_rest.R
\name{mle_rest}
\alias{mle_rest}
\title{Maximu-likelihood estimation of animal density using the REST model}
\usage{
mle_rest(
  npass_effort_data,
  stay_data,
  activity_data,
  col_name_npass = "npass",
  col_name_effort = "effort",
  col_name_stay = "stay",
  col_name_cens = "cens",
  focal_area
)
}
\arguments{
\item{npass_effort_data}{\subsection{}{
}}

\item{stay_data}{\subsection{}{
}}

\item{activity_data}{\subsection{}{
}}

\item{col_name_npass}{\subsection{}{
}}

\item{col_name_effort}{#E#}

\item{col_name_stay}{\subsection{}{
}}

\item{col_name_cens}{\subsection{}{
}}

\item{focal_area}{\subsection{}{
}}
}
\value{
estimated density
}
\description{
Maximu-likelihood estimation of animal density using the REST model
}
\examples{
#' mle_rest(
npass_effort_data = npass_effort_data,
stay_data = stay_data,
activity_data = activity_data,
col_name_npass = "npass",
col_name_effort = "effort",
col_name_stay = "stay",
col_name_cens = "cens",
focal_area)
)
}
