% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_videos.R
\name{select_videos}
\alias{select_videos}
\title{A function to randomly select N_sampled videos such that the number of selected videos per camera station is as evenly distributed as possible.}
\usage{
select_videos(
  detection_data,
  col_name_species = "Species",
  col_name_station = "Station",
  col_name_datetime = "DateTimeCorrected",
  N_sampled,
  Indep_criteria,
  seed = NULL,
  target_species = NULL
)
}
\arguments{
\item{detection_data}{A data frame containing information for individual videos, with six columns:
\itemize{
\item Camera station ID (character)
\item Capture datetime (character, not used in this function)
\item Species name (character)
\item Number of passes through the focal area (numeric, not used in this function)
\item Staying time (seconds) within the focal area for each pass in a video (numeric)
\item Whether the observation of the pass was censored (1 = censored, 0 = observed)
}}

\item{col_name_species}{A string specifying the column name in \code{stay_data} that indicates the species name.}

\item{col_name_station}{A string specifying the column name containing camera station IDs.}

\item{col_name_datetime}{A string specifying the column name that contains datetime information.}

\item{N_sampled}{The total number of videos to be selected.}

\item{Indep_criteria}{A numeric value representing the interval (in minutes) used to determine whether detections are independent. The default is 30 minutes.}

\item{seed}{An integer value for setting the random seed.}

\item{target_species}{A character string specifying the species of interest. Only a single species can be specified.}
}
\value{
A data frame containing the selected videos.
}
\description{
A function to randomly select N_sampled videos such that the number of selected videos per camera station is as evenly distributed as possible.
}
\examples{
select_videos(detection_data = detection_data,
              col_name_species = "Species",
              col_name_station = "Station",
              col_name_datetime = "DateTime",
              N_sampled = 100,
              Indep_criteria = 30,
              target_species = "SP01")
}
