% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_data.R
\docType{data}
\name{station_data}
\alias{station_data}
\title{Virtual dataset of camera trapping detection data}
\format{
\subsection{\code{station_data}}{

A data frame with 2 columns:
\describe{
\item{Station}{station id detecting an animal pass}
\item{x}{a habitat covairate}
}
}
}
\source{
{simulated data}
}
\usage{
station_data
}
\description{
An example of dataset containing info on camera stations
}
\keyword{datasets}
